chore: refactor code to address maintainability concerns
## Changes

### Module Splitting
- Split exec-handler.ts (605 lines → 4 focused modules):
  - exec-validator.ts: Payload validation logic
  - exec-safety.ts: Command safety and env filtering
  - exec-environment.ts: Environment variable utilities
  - exec-handler.ts: Main orchestration (now 135 lines)

- Extract ConnectionPool from ssh-service.ts (506 → 440 lines):
  - connection-pool.ts: Dedicated connection management class

### Code Quality Improvements
- Extract magic numbers in hostname-resolver.ts to named constants:
  - IPV4_MAX_MASK_BITS, IPV4_FULL_MASK, IPV4_OCTET_MAX, etc.
  - IPV6_MAX_MASK_BITS, IPV6_SEGMENT_COUNT, etc.

- Create extractErrorMessage() utility function:
  - Simplifies complex error narrowing in app.ts
  - Reusable across codebase

- Replace console statements with debug logger:
  - service-socket-authentication.ts
  - service-socket-control.ts

### Documentation
- Add comprehensive JSDoc for singleton lifecycle:
  - config.ts: Configuration singleton pattern
  - store.ts: Session store singleton with thread safety notes

## Testing
- ✅ All 604 tests passing
- ✅ TypeScript compilation: 0 errors
- ✅ ESLint: 0 warnings

## Impact
- Improved maintainability through better separation of concerns
- Enhanced code readability with named constants
- Better documentation of singleton patterns
- No functional changes
